Migrate tests to follow Render() code into common directory (#347)
In a previous PR, we migrated the render logic from a CLI command to a
library function. That PR didn't move the tests correspondingly, and
we're doing that now.

This is just code movement without anything interesting.
